+++ src/mask.c
@@ -5,6 +5,7 @@
 #include <stdio.h>
 #include "config.h"
 #include "mask.h"
+#include "mstr.h"
 
 #define PC_TEST 0
 
@@ -20,7 +21,7 @@ int ch;
   }
 }
 
-/***************************************************************************************/
+/*****************************************************************************/
 
 static char *get_pattern_from_pos (char *source, int slen, int pos)
 {
@@ -53,8 +54,7 @@ char *dest, *ptmp;
   return dest;
 }
 
-/***************************************************************************************/
-
+/*****************************************************************************/
 static int is_there_any_sym_in_mask(char *mask, int slen, int pos)
 {
 int i;
@@ -67,8 +67,7 @@ char ch;
   return 0;
 }
 
-/***************************************************************************************/
-
+/*****************************************************************************/
 int mask_match (const char *expr, const char *mask)
 {
 int sourceLen, searchLen;
@@ -175,6 +174,87 @@ char *mexpr, *mmask;
   free (mmask);
   return res;
 }
+
+/*****************************************************************************/
+int masks_match (const char *expr, mask_t *m)
+{
+int i, ret;
+  if ((m->tokens == NULL) || (m->cnt < 1))  {
+    ret = mask_match (expr, "*");
+  }
+  else  {
+    for (i = 0; i < m->cnt; i++)  {
+      ret = mask_match (expr, m->tokens[i]);
+      if (ret)  {
+        break;
+      }
+    }
+  }
+  return ret;
+}
+
+/*****************************************************************************/
+#define ALLOC_STEP  16
+
+static int token_add (char *token, mask_t *m)
+{
+char **p;
+int ret = 0;
+  if (m->cnt >= m->allocated)  {
+    p = (char **)realloc(m->tokens, (m->allocated+ALLOC_STEP)*sizeof(char*));
+    if (p == NULL)  {
+      masks_unprepare (m);
+      ret = 1;
+    }
+    else  {
+      m->tokens = p;
+      m->allocated += ALLOC_STEP;
+      m->tokens[m->cnt++] = trim_string(token);
+      // if leading & trailing spaces are a valid symbols in mask
+      //m->tokens[m->cnt++] = token;
+    }
+  }
+  else  {
+    m->tokens[m->cnt++] = trim_string(token);
+    // if leading & trailing spaces are a valid symbols in mask
+//    m->tokens[m->cnt++] = token;
+  }
+  return ret;
+}
+
+/*****************************************************************************/
+int masks_prepare (const char *mask, mask_t *m)
+{
+char *tok;
+int ret = 0;
+  m->cnt = 0;
+  m->allocated = 0;
+  strncpy (m->bulk, mask, MAX_MASK_LEN-1);
+  m->bulk[MAX_MASK_LEN-1] = 0;
+  m->tokens = NULL;
+
+  tok = strtok (m->bulk, ";,");
+  while (tok != NULL)  {
+    if (!m_is_space_string (tok))  {
+      ret = token_add (tok, m);
+      if (ret)  {
+        break;
+      }
+    }
+    tok = strtok (NULL, ";,");
+  }
+  return ret;
+}
+
+/*****************************************************************************/
+void masks_unprepare (mask_t *m)
+{
+  free (m->tokens);
+  m->tokens = NULL;
+  m->allocated = 0;
+  m->cnt = 0;
+}
+
 
 
 #if PC_TEST